backend-python/devops-data-sentinel/requirements.txt
cryptography
peewee
pyyaml
pymysql
requests
python-json-logger